+++ packages/durabletask-js/test/client-options.spec.ts
@@ -38,6 +38,5 @@ describe("TaskHubGrpcClient", () => {
 
       expect(client).toBeDefined();
     });
-
   });
 });
